ProjectFallDetect/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.5.2
asttokens
astunparse==1.6.3
backcall
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama 
comm 
cycler==0.12.1
debugpy 
decorator
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing 
fastapi==0.115.6
fastapi-cli==0.0.7
flatbuffers==1.12
fonttools==4.55.3
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata
importlib_resources==6.4.5
ipykernel
ipython
jedi
Jinja2==3.1.5
jupyter_client
jupyter_core 
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline
mdurl==0.1.2
mysql-connector-python==9.0.0
nest_asyncio
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging
pandas==2.0.3
parso
pickleshare
pillow==10.4.0
platformdirs
prompt_toolkit 
protobuf==3.19.6
psutil
pure_eval
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments
pyparsing==3.1.4
python-dateutil
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
seaborn==0.13.2
shellingham==1.5.4
six
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data 
starlette==0.41.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tornado
traitlets
typer==0.15.1
typing_extensions 
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
watchfiles==0.24.0
wcwidth 
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
